tests/wpt/tests/css/css-conditional/container-queries/container-selection-unknown-features.html

<!DOCTYPE html>
<title>@container: selection using name and implicit selection</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #container {
    container-type: size;
    width: 200px;
    --foo: bar;
  }
  @container (width > 0px) {
    div { color: green }
  }
  @container (width > 0px) or (foo: bar) {
    #t1 { color: red }
  }
  @container (width > 0px) or foo(bar) {
    #t2 { color: red }
  }
  @container style(--foo: bar) or (foo: bar) {
    #t3 { color: red }
  }
</style>
<div id="container">
  <div id="t1">Green</div>
  <div id="t2">Green</div>
  <div id="t3">Green</div>
  <div id="t4">Green</div>
</div>
<script>
  setup(() => assert_implements_container_queries());

  const green = "rgb(0, 128, 0)";
  test(() => { assert_equals(getComputedStyle(t1).color, green); }, "width query with (foo: bar)");
  test(() => { assert_equals(getComputedStyle(t2).color, green); }, "width query with foo(bar)");
  test(() => { assert_equals(getComputedStyle(t3).color, green); }, "style query with (foo: bar)");
</script>
